:wrench: Fix routing & link
