+++ pages/about.vue
