+++ pages/index.vue
@@ -1,92 +1,35 @@
 <template>
-  <v-layout
-    column
-    justify-center
-    align-center
-  >
-    <v-flex
-      xs12
-      sm8
-      md6
-    >
-      <div class="text-center">
-        <logo />
-        <vuetify-logo />
-      </div>
-      <v-card>
-        <v-card-title class="headline">
-          Welcome to the Vuetify + Nuxt.js template
-        </v-card-title>
-        <v-card-text>
-          <p>Vuetify is a progressive Material Design component framework for Vue.js. It was designed to empower developers to create amazing applications.</p>
-          <p>
-            For more information on Vuetify, check out the <a
-              href="https://vuetifyjs.com"
-              target="_blank"
-            >
-              documentation
-            </a>.
-          </p>
-          <p>
-            If you have questions, please join the official <a
-              href="https://chat.vuetifyjs.com/"
-              target="_blank"
-              title="chat"
-            >
-              discord
-            </a>.
-          </p>
-          <p>
-            Find a bug? Report it on the github <a
-              href="https://github.com/vuetifyjs/vuetify/issues"
-              target="_blank"
-              title="contribute"
-            >
-              issue board
-            </a>.
-          </p>
-          <p>Thank you for developing with Vuetify and I look forward to bringing more exciting features in the future.</p>
-          <div class="text-xs-right">
-            <em><small>&mdash; John Leider</small></em>
-          </div>
-          <hr class="my-3">
-          <a
-            href="https://nuxtjs.org/"
-            target="_blank"
-          >
-            Nuxt Documentation
-          </a>
-          <br>
-          <a
-            href="https://github.com/nuxt/nuxt.js"
-            target="_blank"
-          >
-            Nuxt GitHub
-          </a>
-        </v-card-text>
-        <v-card-actions>
-          <v-spacer />
-          <v-btn
-            color="primary"
-            nuxt
-            to="/inspire"
-          >
-            Continue
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-flex>
-  </v-layout>
+  <v-row>
+    <v-col xs12 sm6 md4>
+      <number-display :title="'本日の感染者'" :number="data.patients ? data.patients.length : 0" :unit="'人'">
+      </number-display>
+    </v-col>
+    <v-col xs12 sm6 md4>
+      <number-display :title="'test'" :number=123 :unit="'人'">
+      </number-display>
+    </v-col>
+    <v-col xs12 sm6 md4>
+      <number-display :title="'test'" :number=123 :unit="'人'">
+      </number-display>
+    </v-col>
+    <v-col xs12 sm6 md4>
+      <number-display :title="'test'" :number=123 :unit="'人'">
+      </number-display>
+    </v-col>
+  </v-row>
 </template>
 
 <script>
-import Logo from '~/components/Logo.vue'
-import VuetifyLogo from '~/components/VuetifyLogo.vue'
+import NumberDisplay from '@/components/NumberDisplay.vue'
+import { SHEET_URL } from '@/constants.js'
 
 export default {
   components: {
-    Logo,
-    VuetifyLogo
-  }
+    NumberDisplay
+  },
+  async asyncData ({ $axios }) {
+    const data = await $axios.$get(SHEET_URL)
+    return { data }
+  },
 }
 </script>
